PR tree-optimization/78913 - Probably misleading error reported by -Wformat-length PR middle-end/77708 - -Wformat-length %s warns for snprintf
gcc/ChangeLog:

	PR middle-end/77708
	* doc/invoke.texi (Warning Options): Document -Wformat-truncation.
	* gimple-ssa-sprintf.c (call_info::reval_used, call_info::warnopt):
	New member functions.
	(format_directive): Used them.
	(add_bytes): Same.
	(pass_sprintf_length::handle_gimple_call): Same.
	* graphite-sese-to-poly.c (tree_int_to_gmp): Increase buffer size
	to avoid truncation for any argument.
	(extract_affine_mul): Same.
	* tree.c (get_file_function_name): Same.

gcc/c-family/ChangeLog:

	PR middle-end/77708
	* c.opt (-Wformat-truncation): New option.

gcc/fortran/ChangeLog:

	PR tree-optimization/78913
	PR middle-end/77708
	* trans-common.c (build_equiv_decl): Increase buffer size to avoid
	truncation for any argument.
	* trans-types.c (gfc_build_logical_type): Same.

gcc/testsuite/ChangeLog:

	PR middle-end/77708
	* gcc.dg/tree-ssa/builtin-snprintf-warn-1.c: New test.
	* gcc.dg/tree-ssa/builtin-snprintf-warn-2.c: New test.
	* gcc.dg/tree-ssa/builtin-sprintf-warn-6.c: XFAIL test cases failing
	due to bug 78969.
	* gcc.dg/format/pr78569.c: Adjust.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@244210 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/tree-ssa/builtin-sprintf-warn-6.c
@@ -51,7 +51,8 @@ void fuint (unsigned j, char *p)
 {
   if (j > 999)
     return;
-  snprintf (p, 4, "%3u", j);
+
+  snprintf (p, 4, "%3u", j);   /* { dg-bogus "may be truncated" "unsigned int" { xfail *-*-* } } */
 }
 
 void fint (int j, char *p)
@@ -61,8 +62,7 @@ void fint (int j, char *p)
   if (k > 999)
     return;
 
-  /* Range info isn't available here.  */
-  snprintf (p, 4, "%3u", k);
+  snprintf (p, 4, "%3u", k);   /* { dg-bogus "may be truncated" "signed int" { xfail *-*-* } } */
 }
 
 void fulong (unsigned long j, char *p)
@@ -70,8 +70,7 @@ void fulong (unsigned long j, char *p)
   if (j > 999)
     return;
 
-  /* Range info isn't available here.  */
-  snprintf (p, 4, "%3lu", j);
+  snprintf (p, 4, "%3lu", j);   /* { dg-bogus "may be truncated" "unsigned long" { xfail *-*-* } } */
 }
 
 void flong (long j, char *p)
@@ -81,8 +80,7 @@ void flong (long j, char *p)
   if (k > 999)
     return;
 
-  /* Range info isn't available here.  */
-  snprintf (p, 4, "%3lu", k);
+  snprintf (p, 4, "%3lu", k);   /* { dg-bogus "may be truncated" "signed long" { xfail *-*-* } } */
 }
 
 void fullong (unsigned long long j, char *p)
@@ -90,18 +88,17 @@ void fullong (unsigned long long j, char *p)
   if (j > 999)
     return;
 
-  /* Range info isn't available here.  */
-  snprintf (p, 4, "%3llu", j);
+  snprintf (p, 4, "%3llu", j);   /* { dg-bogus "may be truncated" "signed long" { xfail *-*-* } } */
 }
 
-void fllong (long j, char *p)
+void fllong (long long j, char *p)
 {
   const unsigned long long k = (unsigned long long) j;
 
   if (k > 999)
     return;
 
-  snprintf (p, 4, "%3llu", k);
+  snprintf (p, 4, "%3llu", k);   /* { dg-bogus "may be truncated" "unsigned long long" { xfail *-*-* } } */
 }
 
 /* { dg-final { scan-tree-dump-not "abort" "optimized" } } */
